pyproject.toml
[tool.poetry]
name = "config_parser"
version = "0.1.0"
description = "Parses cli imports and creates Config pydantic object based on provided values"
authors = ["Tesla2000 <fratajczak124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.1.0"
toml = "^0.10.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
pydantic = {extras = ["mypy"], version = "^2.11.3"}

[tool.mypy]
plugins = ["pydantic.mypy"]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
config_parser = "config_parser:main"
